Move generic filesystem operations to their own module
Turns out we'll need those for posix systems as well. This is just a
rename/move and contains no behavior change whatsoever.

I only moved generic functions. win32fs_multiuser_windows still contains
`grantFullControl` and its friends. I decided to not move those into
another file in this commit for the sake of reviewability but  now that
file name kinda lies, it's not about fs on win32 anymore but just
granting full access to a directory.
